CodeExplanation Gen AI Tool/requirements.txt
streamlit
requests
huggingface-hub
python-dotenv